Move OSutils and netutils to pkg
