+++ pkg/sshutils/testsshserver/TestSshServer.go
@@ -15,7 +15,7 @@ import (
 	"time"
 
 	"github.com/asciich/asciichgolangpublic/logging"
-	"github.com/asciich/asciichgolangpublic/netutils"
+	"github.com/asciich/asciichgolangpublic/pkg/netutils"
 	"github.com/asciich/asciichgolangpublic/pkg/contextutils"
 	"github.com/asciich/asciichgolangpublic/tracederrors"
 	"golang.org/x/crypto/ssh"
